third_party/WebKit/LayoutTests/fast/dom/script-styled-size-expected.txt

Test that script can access its own renderer when executing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS script.offsetWidth > 0 is true
PASS script.offsetWidth > 0 is true
PASS successfullyParsed is true

TEST COMPLETE

shouldBeTrue("script.offsetWidth > 0")
